requirements.txt
#eun 
#pip install -r requirements.txt
opencv-python
pytesseract
python-dotenv
